+++ kotori/daq/application/composite.py
@@ -45,7 +45,7 @@ class CompositeApplication(RootService):
             service = service_factory(
                 channel  = self.channel,
                 strategy = strategy_factory(settings=global_settings),
-                graphing = graphing_factory(settings=global_settings)
+                graphing = graphing_factory(settings=global_settings, channel=self.channel)
                 )
 
             # Register service component with its container