+++ CHANGES.rst
@@ -31,6 +31,7 @@ in progress
 - [grafana] Use topology information for deriving unique dashboard id from
 - [grafana] Add mechanisms for taming the dashboard refresh interval
 - [core] Improve component loading robustness
+- [grafana] Refactoring. Robustness. Run GrafanaManager as service. Add worker service for taming dashboard refresh intervals.
 
 
 .. _kotori-0.20.1: